Tool switchovers (#661)
* Setup UV package installation, environment creation, and CI CD.
* Definitively fix bot timestamps
All timestamps set to be integers. because we don't need the precision. Convert benchmark series to readable dates.
Including fixing bot filtering by start/end date timestamps.

* Chart orderlines fixes
* Fix order status querying
* Fix liquidation validation errors
* Replace get data notifications with spinner, drastically reducing number of toasts

+++ api/deals/base.py
@@ -155,7 +155,7 @@ class BaseDeal(OrderController):
         """
         all_orders = self.get_all_orders(
             self.active_bot.pair,
-            start_time=float(self.active_bot.deal.opening_timestamp),
+            start_time=int(self.active_bot.deal.opening_timestamp),
         )
         for order in all_orders:
             if (
